chore: move middleware to hangfire project, take out thread slot names to common project, move filter registration to UseHangfire extension
