+++ src/VirtoCommerce.Platform.Core/Common/ThreadSlotNames.cs
@@ -1,0 +1,7 @@
+namespace VirtoCommerce.Platform.Core.Common
+{
+    public static class ThreadSlotNames
+    {
+        public const string USER_NAME = "UserName";
+    }
+}
