+++ src/VirtoCommerce.Platform.Hangfire/ApplicationBuilderExtensions.cs
@@ -1,6 +1,7 @@
 using Hangfire;
 using Hangfire.SqlServer;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,6 +9,7 @@ using Microsoft.Extensions.Options;
 using VirtoCommerce.Platform.Core.Bus;
 using VirtoCommerce.Platform.Core.Settings;
 using VirtoCommerce.Platform.Core.Settings.Events;
+using VirtoCommerce.Platform.Hangfire.Middleware;
 
 namespace VirtoCommerce.Platform.Hangfire
 {
@@ -22,7 +24,7 @@ namespace VirtoCommerce.Platform.Hangfire
             // This is an important workaround of Hangfire initialization issues
             // The standard database schema initialization way described at the page https://docs.hangfire.io/en/latest/configuration/using-sql-server.html works on an existing database only.
             // Therefore we create SqlServerStorage for Hangfire manually here.
-            // This way we ensure Hangfire schema will be applied to storage AFTER platform database creation.            
+            // This way we ensure Hangfire schema will be applied to storage AFTER platform database creation.
             var hangfireOptions = appBuilder.ApplicationServices.GetRequiredService<IOptions<HangfireOptions>>().Value;
             if (hangfireOptions.JobStorageType == HangfireJobStorageType.SqlServer)
             {
@@ -39,6 +41,10 @@ namespace VirtoCommerce.Platform.Hangfire
             var recurringJobManager = appBuilder.ApplicationServices.GetService<IRecurringJobManager>();
             var settingsManager = appBuilder.ApplicationServices.GetService<ISettingsManager>();
             inProcessBus.RegisterHandler<ObjectSettingChangedEvent>(async (message, token) => await recurringJobManager.HandleSettingChangeAsync(settingsManager, message));
+
+            // Add Hangfire filters/middlewares
+            var contextAccessor = appBuilder.ApplicationServices.GetRequiredService<IHttpContextAccessor>();
+            GlobalJobFilters.Filters.Add(new HangfireUserContextMiddleware(contextAccessor));
 
             return appBuilder;
         }
