+++ src/VirtoCommerce.Platform.Hangfire/Middleware/HangfireUserContextMiddleware.cs
@@ -2,17 +2,18 @@ using System.Threading;
 using Hangfire.Client;
 using Hangfire.Server;
 using Microsoft.AspNetCore.Http;
-using VirtoCommerce.Platform.Security;
+using static VirtoCommerce.Platform.Core.Common.ThreadSlotNames;
 
-namespace VirtoCommerce.Platform.Web.Middleware.Hangfire
+namespace VirtoCommerce.Platform.Hangfire.Middleware
 {
+    /// <summary>
+    /// This class allow to process all hangifre jobs to add user name from identity and save it to
+    /// the Thread after job is perfoming to achieve getting access to user name in background tasks
+    /// </summary>
     public class HangfireUserContextMiddleware : IClientFilter, IServerFilter
     {
         private readonly IHttpContextAccessor _contextAccessor;
 
-        /// <summary>
-        /// Ctor
-        /// </summary>
         public HangfireUserContextMiddleware(IHttpContextAccessor contextAccessor) => _contextAccessor = contextAccessor;
 
         private string ContextUserName => _contextAccessor?.HttpContext?.User?.Identity?.Name;
@@ -24,7 +25,7 @@ namespace VirtoCommerce.Platform.Web.Middleware.Hangfire
             if (ContextUserName is null)
                 return;
 
-            filterContext.SetJobParameter(HttpContextUserResolver.USER_NAME_THREAD_SLOT_NAME, ContextUserName);
+            filterContext.SetJobParameter(USER_NAME, ContextUserName);
         }
 
         public void OnCreated(CreatedContext filterContext)
@@ -38,9 +39,9 @@ namespace VirtoCommerce.Platform.Web.Middleware.Hangfire
 
         public void OnPerforming(PerformingContext filterContext)
         {
-            var userName = filterContext.GetJobParameter<string>(HttpContextUserResolver.USER_NAME_THREAD_SLOT_NAME);
+            var userName = filterContext.GetJobParameter<string>(USER_NAME);
 
-            Thread.SetData(Thread.GetNamedDataSlot(HttpContextUserResolver.USER_NAME_THREAD_SLOT_NAME), userName);
+            Thread.SetData(Thread.GetNamedDataSlot(USER_NAME), userName);
         }
 
         public void OnPerformed(PerformedContext filterContext)
