+++ src/VirtoCommerce.Platform.Security/HttpContextUserResolver.cs
@@ -1,5 +1,6 @@
 using System.Threading;
 using Microsoft.AspNetCore.Http;
+using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Security;
 
 namespace VirtoCommerce.Platform.Security
@@ -8,8 +9,6 @@ namespace VirtoCommerce.Platform.Security
     {
         private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public const string USER_NAME_THREAD_SLOT_NAME = "UserName";
-
         public HttpContextUserResolver(IHttpContextAccessor httpContextAccessor)
         {
             _httpContextAccessor = httpContextAccessor;
@@ -17,7 +16,7 @@ namespace VirtoCommerce.Platform.Security
 
         public string GetCurrentUserName()
         {
-            var result = Thread.GetData(Thread.GetNamedDataSlot(USER_NAME_THREAD_SLOT_NAME)) as string ?? "unknown";
+            var result = Thread.GetData(Thread.GetNamedDataSlot(ThreadSlotNames.USER_NAME)) as string ?? "unknown";
 
             var context = _httpContextAccessor.HttpContext;
             if (context != null && context.Request != null && context.User != null)
